워커 쓰레드 Execute()함수 인자 자료형 변경
64비트 환경에서 64비트지만, 실제로는 32비트만 써서 DWORD로 변경

+++ ServerCommon_LIB/Utility/private/Thread/ThWorkerJob.cpp
@@ -22,7 +22,7 @@ ThWorkerJob::ThWorkerJob(WokerPtr&& workItem, WORKER_TYPE type)
   ZeroMemory(this, sizeof(OVERLAPPED));
 }
 
-void ThWorkerJob::operator()(const DWORD ioByte, const uint64_t errorCode) {
+void ThWorkerJob::operator()(const DWORD ioByte, const DWORD errorCode) {
   m_workItem->Execute(this, ioByte, errorCode);
   // ZeroMemory(this, sizeof(OVERLAPPED));
 }
