+++ ServerCommon_LIB/Utility/private/Thread/ThreadPool.cpp
@@ -31,7 +31,8 @@ void ThreadPool::RunningThread(std::stop_token stopToken) {
       }
 
       // Internal은 에러 코드
-      workerJob(overlappedEntry[i].dwNumberOfBytesTransferred, overlappedEntry[i].Internal);
+      // 64비트 자료형이지만, 실제로는 32비트만 사용
+      workerJob(overlappedEntry[i].dwNumberOfBytesTransferred, static_cast<DWORD>(overlappedEntry[i].Internal));
     }
     if (stopToken.stop_requested()) {
       return;
